% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_config.R
\name{get_builtin_scenarios}
\alias{get_builtin_scenarios}
\title{Get Built-in Scenario Names}
\usage{
get_builtin_scenarios()
}
\value{
Character vector of scenario names.
}
\description{
Returns the names of available built-in TTIQ scenarios.
}
\examples{
get_builtin_scenarios()

}
