% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gi_dist_cdf.R
\name{gi.dist.cdf}
\alias{gi.dist.cdf}
\title{Generation Interval Cumulative Distribution Function}
\usage{
gi.dist.cdf(q, meanlog = 1.376, sdlog = 0.567)
}
\arguments{
\item{q}{Numeric vector. Quantiles (time values in days).}

\item{meanlog}{Numeric. Mean of the distribution on the log scale.
Default is 1.376 (COVID-19 estimate).}

\item{sdlog}{Numeric. Standard deviation of the distribution on the log scale.
Default is 0.567 (COVID-19 estimate).}
}
\value{
Numeric vector of cumulative probabilities.
}
\description{
Returns the CDF of the lognormal generation interval distribution.
}
\details{
Used to calculate the proportion of secondary transmissions that would have
occurred by a given time point. This is used in calculating infection potential.
}
\examples{
# Probability of transmission within 5 days (COVID-19)
gi.dist.cdf(5)

# Probability of transmission within 3 days (Influenza)
gi.dist.cdf(3, meanlog = 0.91, sdlog = 0.52)

}
