% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nsw.ci.cc.notify.dat}
\alias{nsw.ci.cc.notify.dat}
\title{Close Contact Notification Delay - NSW Case-Initiated Scenario}
\format{
A data frame with 16 rows and 2 columns:
\describe{
\item{delay}{Numeric. Days (0-15)}
\item{cum_prob}{Numeric. Cumulative probability (0-1)}
}
}
\source{
Synthetic data from exponential CDF
}
\usage{
nsw.ci.cc.notify.dat
}
\description{
Cumulative distribution function for time to notify and quarantine close
contacts under NSW baseline conditions.
}
\details{
NSW notification delays with median ~1.5 days.
}
\keyword{datasets}
