% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{optimal.cc.notify.dat}
\alias{optimal.cc.notify.dat}
\title{Close Contact Notification Delay - Optimal Scenario}
\format{
A data frame with 16 rows and 2 columns:
\describe{
\item{delay}{Numeric. Days from interview to contact notification (0-15)}
\item{cum_prob}{Numeric. Cumulative probability (0-1)}
}
}
\source{
Synthetic data from exponential CDF
}
\usage{
optimal.cc.notify.dat
}
\description{
Cumulative distribution function for time to notify and quarantine close
contacts under optimal conditions.
}
\details{
Fast contact notification with median ~0.8 days.
}
\keyword{datasets}
