% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samp_hh_size.R
\name{samp.hh.size}
\alias{samp.hh.size}
\title{Sample Household Sizes}
\usage{
samp.hh.size(n, region = NULL, hh_probs = NULL)
}
\arguments{
\item{n}{Integer. Number of samples to draw.}

\item{region}{Character. Region name for built-in household size distribution.
For Australian states, see \code{\link{abbreviate_states}} for valid values.
Ignored if \code{hh_probs} is provided.}

\item{hh_probs}{Numeric vector. Custom household size probabilities for sizes 1-8+.
Must sum to 1 and have length 8. If provided, \code{region} is ignored.}
}
\value{
Integer vector of length n containing household sizes (1-8).
}
\description{
Draws n household sizes from a specified distribution using inverse transform sampling.
}
\details{
If \code{hh_probs} is provided, it should be a vector of 8 probabilities
corresponding to household sizes 1, 2, 3, 4, 5, 6, 7, and 8+ persons.
}
\examples{
# Using built-in Australian data
samp.hh.size(100, region = "NSW")

# Using custom UK-like distribution
uk_probs <- c(0.28, 0.35, 0.18, 0.12, 0.05, 0.015, 0.004, 0.001)
samp.hh.size(100, hh_probs = uk_probs)

}
