% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_config.R
\name{sample_from_dist}
\alias{sample_from_dist}
\title{Sample from a Distribution Specification}
\usage{
sample_from_dist(n, dist_spec)
}
\arguments{
\item{n}{Integer. Number of samples.}

\item{dist_spec}{List. Distribution specification with type and parameters.}
}
\value{
Numeric vector of samples.
}
\description{
Draws n samples from a distribution specified as a list.
}
\keyword{internal}
