% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_named_vector.R
\name{to.named.vector}
\alias{to.named.vector}
\title{Convert Data Frame to Named Vector}
\usage{
to.named.vector(df)
}
\arguments{
\item{df}{Data frame with two columns. First column becomes names,
second column becomes values.}
}
\value{
Named numeric vector.
}
\description{
Converts a two-column data frame to a named vector.
}
\keyword{internal}
